Work on templates, loggin in and sign up works
Added templates for order, product and sign up. Added functionality for
these aforementioned parts. Added `utils.py` with the `is_mail()`
function for regex checking for email addresses. Other templates were
changed or implemented.

More work needs to be done the navbar and the 'html and css' part of
things. More order functionality needs to be implemented as well as
using the catalogue for quickly adding new orders and products to the
cart. Admin account needs to be created and implemented, so that new
orders can be added easily.